requirements/prod.txt
colorama==0.4.1
GitPython==2.1.11
pandas==0.24.2
path.py==12.0.1
